Add fix for right frame and units
I changed the secondary rotation speed expression from the A frame to
the B frame (for (F)Hamiltonian files). Now it is expressed in B as
explained in the user guide.

Alsi I changed the units of the potential energy, the kinetic energy,
the angular momentum, etc. to be in the right MKS units as explained in
the user guide.